requirements.txt
flask
flask-sqlalchemy
python-dotenv
openai